Update new_game_data_pull.py
So i'll first just try to describe the changes I was suggesting on reddit.  I'm not quite following all your pandas operations here.

in general, you will I think want to add a new column to your pandas table which is the thing_id from bgg.  This is the number in the url:

https://boardgamegeek.com/boardgame/10865/road-block

the thing_id here is 10865.

I added a function above (get_thing_id_from_url) that works ok for this

Next, you will want to build a list of the new thing ids in the if not new_entries.empty area.  Then batch those in up to 100 at a time to this API for bgg:

https://boardgamegeek.com/xmlapi2/thing?id=1,2,3,4

that call will get thing ids 1, 2, 3 and 4 - it will come back in an xml format.

then it seems like all you really want from that is the name / title and the year.  you can get those similar to the dummy_bgg_call function I put above.  This also has some player count stuff too

I don't have a ton of time right now to dig into this, but hopefully this is a good start for you.  Let me know if you have any questions.